# crates/byos/openapi.yml
openapi: 3.0.3
info:
  title: BYOS Proposal API
  description: |
    The public HTTP API by which sub-solvers submit signed settlement
    proposals to the BYOS solver engine (ADR-0001).

    The API is permissionless but collateral-gated: there is no registration
    or API key. A sub-solver's identity is the address recovered from its
    EIP-712 signatures, and its escrow deposit is the permission.

    ## Conventions (ADR-0005)

    Follows the CoW Protocol solver-engine wire idiom: camelCase JSON,
    256-bit amounts as decimal strings, addresses and order UIDs as
    `0x`-prefixed hex strings.

    ## Authentication

    Every endpoint except `POST /proposals` and `GET /healthz` requires an
    EIP-712 signature in the `X-Signature` header, hex-encoded (65 bytes).
    The signed message differs by operation:

    - **Reads** (`GET`) sign the `ReadAuth` struct — `ReadAuth(uint256
      version)` with `version = 1`. It is a bearer token: signed once, sent
      on every request, no nonce or expiry (ADR-0011).
    - **Cancellation** (`DELETE`) signs the `CancelProposal` struct —
      `CancelProposal(uint256 proposalId)` — so each cancellation is
      authorized individually.

    All signatures share one EIP-712 domain: `name = "BYOS"`,
    `version = "0.1"`, `chainId` of the target chain, and
    `verifyingContract` set to the TrampolineFactory address.

    Reads and cancellations are owner-scoped (ADR-0011): the recovered
...

    ## Proposal lifecycle

...
    the verdict.

...
servers:
...
paths:
...
  /proposal/{id}:
    parameters:
      - $ref: "#/components/parameters/ProposalIdPath"
    get:
      operationId: getProposal
      summary: Get one of the caller's proposals.
      description: |
        Returns the proposal's amounts, status, and, depending on `status`,
        the machine-readable `rejectionReason` (`rejected`) or the
        `settlementTxHash` (`settled`/`settleFailed`/`penalized`). The
        submitted `interactions`, `nonce`, and `signature` are not echoed
        back.
        Owner-scoped: the proposal is only returned if the `X-Signature`
        signer is its sub-solver.
      security:
        - readAuth: []
      responses:
        "200":
          description: The caller's proposal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProposalResponse"
        "400":
          description: Missing or invalid `X-Signature` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: |
            No such proposal — or a proposal owned by another sub-solver;
            the two cases are indistinguishable on the wire (ADR-0011).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: cancelProposal
      summary: Cancel a live proposal.
      description: |
        Requires an EIP-712 `CancelProposal` signature over the proposal id
        in the `X-Signature` header. The recovered signer must be the
        proposal's sub-solver; a non-owner gets the same `404` as a genuine
        miss (ADR-0011). Only live proposals (`submitted` or `active`) can
        be cancelled.
      security:
        - cancelAuth: []
      responses:
        "204":
          description: Proposal cancelled.
        "400":
          description: Missing or invalid `X-Signature` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: |
            No such proposal — or a proposal owned by another sub-solver;
            the two cases are indistinguishable on the wire (ADR-0011).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: |
            The proposal is not cancellable: `executing` (a settlement
            built on it is in flight, ADR-0013) or already terminal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    readAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: |
        EIP-712 signature over the `ReadAuth` struct (`version = 1`),
        hex-encoded (65 bytes). A bearer token: the recovered signer scopes
        what the caller may read (ADR-0011).
    cancelAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: |
        EIP-712 signature over the `CancelProposal` struct for the proposal
        id being cancelled, hex-encoded (65 bytes).
  parameters:
    ProposalIdPath:
      name: id
      in: path
      required: true
      description: Server-assigned proposal id.
      schema:
        $ref: "#/components/schemas/ProposalId"
  schemas:
    Address:
      description: An Ethereum address, 20 bytes as a `0x`-prefixed hex string.
      type: string
      pattern: "^0x[0-9a-fA-F]{40}$"
      example: "0x9008d19f58aabd9ed0d60971565aa8510560ab41"
    OrderUid:
      description: |
        A 56-byte CoW Protocol order UID (32-byte order hash + 20-byte owner
        + 4-byte validTo) as a hex string. The `0x` prefix is optional on
        input; responses always include it.
      type: string
      pattern: "^(0x)?[0-9a-fA-F]{112}$"
    U256:
      description: A 256-bit unsigned integer as a decimal string.
      type: string
      pattern: "^[0-9]+$"
      example: "1000000000000000000"
    UnixTimestamp:
      description: |
        A Unix timestamp in seconds, as a decimal string (256-bit range on
        the wire; the signed `ProposalData.validUntil` is a `uint256`).
      type: string
      pattern: "^[0-9]+$"
      example: "1753650000"
...
    ProposalId:
      description: Server-assigned proposal id (unsigned 64-bit integer).
      type: integer
      format: int64
      minimum: 0
      example: 42
    ProposalStatus:
      description: |
        Lifecycle state of a proposal (ADR-0013):

        - `submitted` — signature verified, awaiting background validation.
        - `active` — passed gatekeeping; eligible for `/solve` selection.
        - `rejected` — failed a gatekeeping rule; see `rejectionReason`.
        - `expired` — `validUntil` passed.
        - `executing` — won an auction; the settlement is being submitted.
          Exits via the driver's outcome or a timeout back to `active`.
        - `settled` — the settlement landed on-chain; see
          `settlementTxHash`.
        - `settleFailed` — the settlement reverted on-chain (see
          `settlementTxHash`); the Track A escrow debit follows.
        - `penalized` — the Track A debit landed.
        - `simFailed` — the settlement simulation reverted.
        - `cancelled` — cancelled by a signed `DELETE`.

        All states except `submitted`, `active`, `executing`, and
        `settleFailed` are terminal.
      type: string
      enum:
        - submitted
        - active
        - rejected
        - expired
        - executing
        - settled
        - settleFailed
        - penalized
        - simFailed
        - cancelled
    RejectionReason:
      description: |
        Why the background validator rejected a proposal (ADR-0007):

        - `InsufficientEscrow` — the sub-solver's escrow balance is below
          the required minimum. Fix: deposit escrow and resubmit.
        - `UnsupportedOrder` — the order is outside the simulation envelope
          (hooks, partial fill, non-erc20 balances — ADR-0012).
        - `AmountMismatch` — the proposal's fill-or-kill amount differs from
          the order's (sell amount for sell orders, buy amount for buy
          orders).
        - `OrderNotFound` — the orderbook does not know the proposal's
          order UID.
        - `Unprofitable` — the first simulation scored the proposal at or
          below the minimum (`score = surplus + fee - gas`, ADR-0002), so it
          could never win an auction (ADR-0013). Fix: improve the price or
          resubmit when gas is cheaper.
      type: string
      enum:
        - InsufficientEscrow
        - UnsupportedOrder
        - AmountMismatch
        - OrderNotFound
        - Unprofitable
...
    GetProposalResponse:
      type: object
      required:
        - id
        - subSolver
        - orderUid
        - sellAmount
        - buyAmount
        - validUntil
        - status
      properties:
        id:
          $ref: "#/components/schemas/ProposalId"
        subSolver:
          description: The sub-solver address recovered from the proposal signature.
          allOf:
            - $ref: "#/components/schemas/Address"
        orderUid:
          $ref: "#/components/schemas/OrderUid"
        sellAmount:
          $ref: "#/components/schemas/U256"
        buyAmount:
          $ref: "#/components/schemas/U256"
        validUntil:
          $ref: "#/components/schemas/UnixTimestamp"
        status:
          $ref: "#/components/schemas/ProposalStatus"
        rejectionReason:
          description: Only present when `status` is `rejected`.
          allOf:
            - $ref: "#/components/schemas/RejectionReason"
        settlementTxHash:
          description: |
            Only present on settlement outcomes: the landed transaction for
            `settled`, the reverted one for `settleFailed`/`penalized`
            (ADR-0013).
          type: string
          pattern: "^0x[0-9a-f]{64}$"
          example: "0x2d6cea6413a3b2e37a1a5c1780e17dee056bd63c2fd0fd60d75bfa9ea38e2e28"
...
    Error:
      description: |
        JSON error body (ADR-0007). `kind` is the machine-readable
        catalogue; `description` is human-readable and may vary.
      type: object
      required:
        - kind
        - description
      properties:
        kind:
          type: string
          enum:
            - InvalidSignature
            - SignatureRecoveryFailed
            - InsufficientEscrow
            - ProposalExpired
            - ProposalLifetimeExceeded
            - ProposalNotFound
            - ProposalNotCancellable
            - BadRequest
            - Internal
        description:
          type: string
          example: "Proposal validUntil is in the past"